MSSQL changes Added hashcode function in faciliofield In UpdateRecordBuilder, combined update query has been separated All Generic builder are compatible with MSSQL Made SQLScriptRunner to execute all the query as single query

+++ src/main/java/com/facilio/bmsconsole/util/NotificationAPI.java
@@ -135,7 +135,7 @@ public class NotificationAPI {
 		
 		List<Map<String, Object>> props = selectBuilder.get();
 		if(props != null && !props.isEmpty()) {
-			Long count = (Long) props.get(0).get("unseen");
+			Integer count = (Integer) props.get(0).get("unseen");
 			return count.intValue();
 		}
 		return 0;
@@ -160,7 +160,7 @@ public class NotificationAPI {
 		
 		List<Map<String, Object>> props = selectBuilder.get();
 		if(props != null && !props.isEmpty()) {
-			Long count = (Long) props.get(0).get("unread");
+			Integer count = (Integer) props.get(0).get("unread");
 			return count.intValue();
 		}
 		return 0;
